Test function parse_individual_games
