+++ vivid_chart.py
@@ -7,37 +7,7 @@
 import plotly.graph_objects as go
 import chess.pgn, time, random, collections
 from collections import Counter
-
-def create_game_list(pgn, depth):
-    ''' we have one massive pgn that we convert to a list of normal game pgns '''
-    game_list = []
-    while True:
-        game = chess.pgn.read_game(pgn)
-        if game is None: # continue until exhausted all games
-            break
-        else:
-            if len(list(game.mainline_moves())) > depth-1: # for graph visualization
-                game_list.append(game)
-    return game_list
-
-def parse_individual_games(pgn, depth):
-    ''' convert each game in the list to SAN format '''
-    full_list = []
-    game_list = create_game_list(pgn, depth)
-    for game in game_list:
-        small_list = [] # essentially only that game's list
-        board = game.board()
-        i = 0
-        for move in game.mainline_moves():
-            if i<depth: # only count moves to needed depth
-                i+=1
-                small_list.append(chess.Board.san(board, move = move))
-                board.push(move)
-            else:
-                break
-        full_list.append(small_list)
-
-    return full_list
+import game_parser
 
 def form_values(depth):
     ''' create parent, id, labels, and values '''
@@ -102,7 +72,7 @@ def form(ids, labels, parents, values):
 user_input_game_file = input('Which game file should be analyzed? Provide FULL path file. ')
 gammme = open(user_input_game_file)
 user_input_depth = int(input('To what ply depth should we visualize these games? '))
-lst = parse_individual_games(gammme, user_input_depth) # ask for input here
+lst = game_parser.parse_individual_games(gammme, user_input_depth) # ask for input here
 
 ids, labels, parents, values = form_values(user_input_depth)
